Refactorizacion de las 3 capas clientes y productos completo exexception domain

+++ TiendaOnline/domain/src/main/java/mx/com/tiendaonline/compraproducto/model/entity/CompraProducto.java
@@ -16,7 +16,7 @@ public class CompraProducto {
     private CantidadProductos cantidad;
     private PrecioUnitario precioUnitario;
 
-    public CompraProducto(Producto producto, Integer cantidad, BigDecimal precioUnitario) {
+    public CompraProducto(Producto producto, Integer cantidad, Double precioUnitario) {
         this.producto = producto;
         this.cantidad = new CantidadProductos(cantidad);
         this.precioUnitario = new PrecioUnitario(precioUnitario);
